CMS completed

+++ classes/Ijdb/Controllers/Joke.php
@@ -3,6 +3,7 @@
 namespace Ijdb\Controllers;
 
 use DateTime;
+use Ninja\Authentication;
 use Ninja\DatabaseTable;
 
 class Joke
@@ -10,10 +11,11 @@ class Joke
     private $authorsTable;
     private $jokesTable;
 
-    public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable)
+    public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication)
     {
         $this->jokesTable = $jokesTable;
         $this->authorsTable = $authorsTable;
+        $this->authentication = $authentication;
     }
 
     public function lists()
@@ -31,6 +33,7 @@ class Joke
                 'jokedate' => $joke['jokedate'],
                 'authorname' => $author['authorname'],
                 'authoremail' => $author['authoremail'],
+                'authorid' => $author['id'],
             ];
         }
 
@@ -38,9 +41,12 @@ class Joke
 
         $totalJokes = $this->jokesTable->total();
 
+        $author = $this->authentication->getUser();
+
         return ['template' => 'jokes.html.php', 'title' => $title, 'variables' => [
             'totalJokes' => $totalJokes,
             'jokes' => $jokes,
+            'userId' => $author['id'] ?? null,
         ]];
 
     }
@@ -54,6 +60,14 @@ class Joke
 
     public function delete()
     {
+        $author = $this->authentication->getUser();
+
+        $joke = $this->jokesTable->findById($_POST['jokeid']);
+
+        if ($joke['authorid'] != $author['id']) {
+            return;
+        }
+
         $this->jokesTable->delete($_POST['jokeid']);
 
         header('Location: /joke/list');
@@ -61,9 +75,19 @@ class Joke
 
     public function saveEdit()
     {
+        $author = $this->authentication->getUser();
+
+        if (isset($_GET['jokeid'])) {
+            $joke = $this->jokesTable->findById($_GET['jokeid']);
+
+            if ($joke['authorid'] != $author['id']) {
+                return;
+            }
+        }
+
         $joke = $_POST['joke'];
         $joke['jokedate'] = new DateTime();
-        $joke['authorid'] = 1;
+        $joke['authorid'] = $author['id'];
 
         $this->jokesTable->save($joke);
 
@@ -72,13 +96,15 @@ class Joke
 
     public function edit()
     {
+        $author = $this->authentication->getUser();
+
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
 
         $title = 'Edit Joke';
 
-        return ['template' => 'editjoke.html.php', 'title' => $title, 'variables' => ['joke' => $joke ?? null]];
+        return ['template' => 'editjoke.html.php', 'title' => $title, 'variables' => ['joke' => $joke ?? null, 'userId' => $author['id'] ?? null]];
 
     }
 }
